reddit-backend/src/main/java/irerin07/repository/UserRepository.java
package irerin07.repository;

import irerin07.domain.User;
import org.springframework.data.jpa.repository.JpaRepository;
import org.springframework.data.jpa.repository.Modifying;
import org.springframework.data.jpa.repository.Query;
import org.springframework.data.repository.query.Param;
import org.springframework.stereotype.Repository;

@Repository
public interface UserRepository extends JpaRepository<User, Long> {
    
    @Query("SELECT u FROM User u WHERE u.email =:email")
    public User findByEmail(@Param("email") String email);

    @Modifying
    @Query("UPDATE User u SET u.nickname = :nickname WHERE u.id = :id")
    public void updateUserNickName(@Param("nickname")String nickname,
                                   @Param("id") long id);

}
